separate tasks for integration and unit tests
