+++ README.md
@@ -2,3 +2,6 @@
 GitHub Actions school project
 
 # Terminal commands for testing:
+- All tests: ./gradlew test
+- Unit tests: ./gradlew unitTest
+- Integration tests: ./gradlew integrationTest
