--- src/main/java/com/example/singlaslantgithubactions/SinglaSlantGitHubActionsApplication.java
@@ -1,16 +1,0 @@
-package com.example.singlaslantgithubactions;
-
-import com.example.singlaslantgithubactions.Model.CoinFlip;
-import com.example.singlaslantgithubactions.Model.RoundResult;
-import com.example.singlaslantgithubactions.Services.Game;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SinglaSlantGitHubActionsApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SinglaSlantGitHubActionsApplication.class, args);
-    }
-
-}
