+++ src/main/java/com/example/singlaslantindividual/SinglaSlantIndividualApplication.java
@@ -1,0 +1,13 @@
+package com.example.singlaslantindividual;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SinglaSlantIndividualApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SinglaSlantIndividualApplication.class, args);
+    }
+
+}
