+++ src/test/java/com/example/singlaslantindividual/Model/RoundResultTest.java
@@ -1,4 +1,4 @@
-package com.example.singlaslantgithubactions.Model;
+package com.example.singlaslantindividual.Model;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
