+++ src/test/java/com/example/singlaslantindividual/SinglaSlantIndividualApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.singlaslantgithubactions;
+package com.example.singlaslantindividual;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class SinglaSlantGitHubActionsApplicationTests {
+class SinglaSlantIndividualApplicationTests {
 
     @Test
     void contextLoads() {
